Add files via upload
Hello! I will be documenting my submissions for Advent of Code 2022. I am not experienced in Python, but this seemed like an interesting way of learning more about it. These may not be the easiest or the most efficient way of doing it, but it's whatever came to mind when completing the exercises.